+++ lib/src/deps.dart
@@ -1,2 +1,1 @@
 export '../database_repository.dart';
-export 'util/json.dart';